Replace Eric Meyer reset.css with normalize.css
This is admittedly an opinionated change, but it's to reduce the overall quantity, complexity and specificity of CSS that we have to write. By using default browser typography styling and padding, we don't have to specify (and then later override) our own when marking up body content, and we won't need a class for each paragraph and list item, which makes things easier for newbies to markup.