Configure Layer Control for Map
Used the .addTo(map) method on the Riders and Drivers layer when they
are first created with Omnivore. This means they will be on by default
in the map controller. Added 'null' first parameter to
L.control.Layers(null, overlaylabels).addTo(map); to define these as
overlay layers not a basemap. The basemap is a toggle radio button while
overlay layers get check boxes.