Use jQuery $.getJSON() to call API for Drivers
Changed out the variable for jsonDriver to call the live API using
jQuery after successful testing on test API. Wrapped the entire script
in a $().ready function to make sure it only runs after elements are
ready. Next step is to make two calls, one for drivers, one for riders.