Fix styling/semantics issues introduced in #244
A previous PR introduced some sitewide styling which negatively affected a lot of the headings across the site. I've removed this style as we don't want all H3/H4 headings to be centered. I've also left-aligned some text that was centered, as in general left-aligned text is more readable when it is more than a few words.

I've also exchanged headings for p/b tags, because headings should be used semantically to indicate the title of a section, not to influence styling. I've instead added a 'highlight' class for emphasised text.