Replace 2016 countdown widget with special election info
The old countdown bannerbox contained an embeddable countdown widget
from 270towin.com. That widget was designed specifically for the 2016
Presidential election, and we have no way of updating the widget for the
2017  elections.

For the time being, we replace the widget with dates for the NJ and
VA state elections. A new countdown widget will need to be developed.

This commit resolves item 17 in #254.