Update accommodation section in driver sign-up form
RiderAccomodationNotes now exist as both placeholder text and as
footnotes to the <textarea> input box. Before they only existed as
placeholder text, which was difficult for vision-impaired users to see.
Furthermore, it posed a usuability problem as the helper text would
disappear as soon as the user entered any text.

This commit also adds a checkbox for "I can only communicate in a
language other than English".

This commit resolves item 12 in #254.